+++ plugins/org.eclipse.elk.alg.layered/src/org/eclipse/elk/alg/layered/intermediate/LongEdgeSplitter.java
@@ -141,7 +141,7 @@ public final class LongEdgeSplitter implements ILayoutProcessor {
         LPort oldEdgeTarget = edge.getTarget();
         
         // Set thickness of the edge
-        float thickness = edge.getProperty(LayeredOptions.EDGE_THICKNESS);
+        double thickness = edge.getProperty(LayeredOptions.EDGE_THICKNESS);
         if (thickness < 0) {
             thickness = 0;
             edge.setProperty(LayeredOptions.EDGE_THICKNESS, thickness);
